move IListItemsProps to common file add common/index.ts
